Backend/ai-security-gateway/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
pydantic
